test(saas_starter): rewrite auth/tenancy tests as Tier-2 integration
Rewrite removes module-skip gate by moving fork+asyncio-incompatible
workflow execution to Tier-2 where infrastructure is real, per brief
failure-layer #3 + zero-tolerance Rule 4.

The original tests/unit/templates/test_saas_starter_auth.py (1375 LOC)
was parked behind pytestmark.skip because Tier-1 top-imports of
LocalRuntime + WorkflowBuilder + runtime.execute(workflow) against
aiosqlite hung the GH-runner py3.11 worker. Probe verdict
(workspaces/issue-979-workstream-b-parallel/journal/0010-PROBE-b-2e-auth-orphan-map.md):
20 of 20 tests map to existing production targets — zero orphans.

The rewrite at tests/integration/templates/saas_starter/test_auth_integration.py
exercises the production workflows.auth.* + middleware.tenant.*
helpers end-to-end against file-backed SQLite (per the templates/*
carve-out documented in tests/CLAUDE.md:109-122). All 20 tests pass
under the default suite; no mocks; real PyJWT + bcrypt + DataFlow
workflow execution.

Two latent bugs in the original Tier-1 file were caught and fixed during
the rewrite (per zero-tolerance Rule 1, "if you found it, you own it"):

1. test_token_expiration_workflow: used naive datetime.now() for the
   token's exp claim. PyJWT interprets exp as UTC, so in any local
   timezone east of UTC the "expired" token actually pointed to a
   future timestamp and the validator accepted it. Fixed to use
   datetime.now(UTC). The module-level skip masked this regression.

2. LocalRuntime fixture: returned a bare LocalRuntime() which the
   v0.12.0 deprecation warning flags ("Use 'with LocalRuntime() as
   runtime:' pattern"). Switched to yield+context-manager form,
   clearing 11 DeprecationWarning emissions per test run.

Out of scope (deferred to a separate shard): one remaining
DeprecationWarning class surfaces ONLY under strict
-W error::DeprecationWarning mode — the Python 3.12 sqlite3 default
datetime adapter (aiosqlite/core.py:63), triggered by Subscription
model datetime fields in test_model_relationships. Confirmed
pre-existing — the sibling test_subscriptions.py exhibits the same
failure at packages/kailash-dataflow/src/dataflow/features/express.py:590
under strict mode. The fix lives in kailash core or DataFlow's
datetime serialization path (multi-package, multi-call-graph-hop)
and exceeds the B-2e shard budget; the canonical disposition is its
own shard per autonomous-execution.md MUST Rule 1.

Closes the B-2e shard of issue #996 (Workstream-B parallel wave).
